pyproject.toml
[tool.poetry]
name = "pack"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
django-formtools = "^2.4.1"
crispy-bootstrap5 = "^2023.10"
django-bootstrap-icons = "^0.8.6"
django-crispy-forms = "^2.1"
django-bootstrap5 = "^23.3"
django-dbbackup = "^4.0.2"
setuptools = "^68.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
